requirements.txt
aiohttp==3.6.2
aiohttp-jwt==0.6.1
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
idna==2.10
multidict==4.7.6
PyJWT==1.7.1
yarl==1.5.1
